+++ src/components/icon/icons/index.js
@@ -200,6 +200,14 @@ export const StarOutline = (props) => {
     )
 }
 
+export const StarHalfFull = (props) => {
+    return (
+        <svg width={props.size ? props.size : defaultSize} height={props.size ? props.size : defaultSize} viewBox="0 0 24 24">
+            <path fill={props.color ? props.color : defaultColor} d="M12,15.4V6.1L13.71,10.13L18.09,10.5L14.77,13.39L15.76,17.67M22,9.24L14.81,8.63L12,2L9.19,8.63L2,9.24L7.45,13.97L5.82,21L12,17.27L18.18,21L16.54,13.97L22,9.24Z" />
+        </svg>
+    )
+}
+
 export const Smartphone = (props) => {
     return (
         <svg width={props.size ? props.size : defaultSize} height={props.size ? props.size : defaultSize} viewBox="0 0 24 24">