+++ src/components/index.js
@@ -49,6 +49,7 @@ export { default as Header } from './header';
 export { default as Dialog } from './dialog';
 export { default as Timeline } from './timeline';
 export { default as Tabs } from './tabs';
+export { default as Rating } from './rating';
 export { 
     convertToNum, 
     compare,