+++ src/pages/RadioPage.js
@@ -145,7 +145,7 @@ const RadioPage = () => {
 
     return (
         <div className="rui-page">
-            <div className="rui-page-title">Radio Group</div>
+            <div className="rui-page-title">RadioGroup Component</div>
             <Card outlined title="Horizontal position with label">
                 <RadioGroup 
                     value={city}